pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-drf-pagination"
version = "0.1.2"
description = "Zero-boilerplate pagination helpers for Django REST Framework"
authors = [{ name="EKRAMUL ISLAM SHADIK" }]
dependencies = ["djangorestframework>=3.14"]
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
  "Framework :: Django",
]

[project.urls]
Homepage = "https://pypi.org/project/simple-drf-pagination/"
Source = "https://github.com/shadikhasan/simple-drf-pagination"

[tool.setuptools.packages.find]
where = ["."]
include = ["simple_drf_pagination*"]
